- portrait orientation support - fix non touch keyboard nav issue - vsync support - clip floppy sounds - reset floppy snd - floppy snd works better with turbo / frameskip - debug show sound buffer usage

+++ Projects/Hataroid/src/main/jni/SDL/src/thread/pthread/SDL_syssem.c
@@ -149,7 +149,7 @@ int SDL_SemWaitTimeout(SDL_sem *sem, Uint32 timeout)
 #else
 	end = SDL_GetTicks() + timeout;
 	while ((retval = SDL_SemTryWait(sem)) == SDL_MUTEX_TIMEDOUT) {
-		if (((long)SDL_GetTicks() - (long)end) >= 0) {
+		if ((SDL_GetTicks() - end) >= 0) {
 			break;
 		}
 		SDL_Delay(0);
